mqtt_logger/requirements.txt
paho-mqtt
pymysql
